Rewrite replacer to operate in a single pass.
Currently the replacement algorithm makes many passes over the input
which results in unwanted replacement when the output of one mapping
contains the input of another as a substring. For example,
`{'main.css': 'main-fp1.css', 'main.css.map': 'main.css-fp2.map'}`
rewrites 'main.css.map' to 'main-fp1.css-fp2.map'.

This implements a new replacement algorithm that operates in a single
pass for all paths simultaneously.

One additional semantic change is that absolute and protocol-relative
URLs will no longer have anything prepended. This generalizes the behavior
from sourcemap replacement and I believe is consistent with the intended
usage of this library.