Add support for rewriting map files
Map files have a `file` attribute that contains the filename of the minified Javascript file they describe. Everything else in the map file is a description of the source code, and shouldn't get re-written. This commit adds custom processing for map files so that if `.map` is included in `replaceExtensions` the correct processing will occur.

I did not add `.map` to the default `replaceExtensions` as that is more of an invasive change from a compatibility standpoint than I really want to make without more demand for this functionality and/or confirmation that it works in use cases other than mine.