Fix prepend option in srcset attributes
The problem in rickharrison/broccoli-asset-rev#48 happened because
the logic to avoid prepending assets twice did not take into account
the edge case of srcset, where there are many assets in one match.

This edge case complicates things more than a bit. To simplify the
solution, when a prepend option is found, we use a replacer function
as second parameter.

Closes rickharrison/broccoli-asset-rev#48